requirements.txt
pandas
openpyxl
numpy
PyPDF2
reportlab
Pillow
num2words
pikepdf
pywin32
